# src/main/resources/Documentation/openapi.yaml
openapi: 3.0.1
info:
  title: RideShare User Service
  description: API Documentation for User Service
  version: "2"
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /batches:
    get:
      tags:
      - Batch
      summary: Return all batches
      description: Returns all batches
      operationId: getBatches
      parameters:
      - name: location
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Batch'
    put:
      tags:
      - Batch
      summary: Update specified batch
      description: Updates batch
      operationId: updateBatch
      requestBody:
        description: Batch to update
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/Batch'
        required: true
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Batch'
        "400":
          description: Input was invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
      - Batch
      summary: Create batch
      description: Adds a new batch
      operationId: addBatch
      requestBody:
        description: User to create
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/Batch'
        required: true
      responses:
        "201":
          description: Batch created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Batch'
        "400":
          description: Input was invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /batches/{number}:
    get:
      tags:
      - Batch
      summary: Return specified batch
      description: Returns batch by number
      operationId: getBatchByNumber
      parameters:
      - name: number
        in: path
        description: Id of batch
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Batch'
        "404":
          description: The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
      - Batch
      summary: Delete specified batch
      description: Deletes batch by number
      operationId: deleteBatchByNumber
      parameters:
      - name: number
        in: path
        description: Batch to delete
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Input was invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Error:
      required:
      - code
      - message
      type: object
      properties:
        code:
          type: string
        message:
          type: string
...
    Batch:
      required:
      - batchLocation
      type: object
      properties:
        batchNumber:
          type: integer
          format: int32
          example: 1
        batchLocation:
          type: string
          example: Morgantown
...